requirements.txt
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
facebook-sdk==3.1.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth-httplib2==0.1.0
google-auth==1.35.0
google-auth-oauthlib==0.5.1
googleads==31.0.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
isodate==0.6.1
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.9.2
lxml==4.8.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pymongo==4.1.0
pyodbc==4.0.32
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
xmltodict==0.12.0
zeep==4.1.0
